Add BatchHeader#full_company_identification in parsing
The parser removes a leading 1 in the #company_identification header,
because I had assumed that this would always imply that the company
identification was an EIN (and an earlier version of #parse always
stripped the leading character because I originally setup
the #company_indentification to only use 9 characters with a separate
code designator field).

This seems like the best non-breaking solution for now, although if ever
actually finish the field_types branch, I'll take a different approach
there.